Document model fields type changed

+++ riznica-web/src/main/webapp/senchaWorkspace/extApps/riznica/classic/src/document/model/DocumentModel.js
@@ -4,14 +4,14 @@ Ext.define("riznica.document.model.DocumentModel", {
   fields: [
     {name: 'documentNumber', type: 'string'},
     {name: 'senderDepartment', type: 'auto'},
-    {name: 'senderId', type: 'long'},
+    {name: 'senderId', type: 'int'},
     {name: 'sender', type: 'string'},
-    {name: 'approverId', type: 'long'},
+    {name: 'approverId', type: 'int'},
     {name: 'approver', type: 'string'},
-    {name: 'receiverId', type: 'long'},
+    {name: 'receiverId', type: 'int'},
     {name: 'receiver', type: 'string'},
     {name: 'remark', type: 'string'},
-    {name: 'approverId', type: 'long'},
+    {name: 'approverId', type: 'int'},
     {name: 'documentType', type: 'auto'},
     {name: 'documentStatus', type: 'auto'}
   ]
